% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelTrain.R
\name{bench_mark}
\alias{bench_mark}
\title{Compare classification models with small samples.}
\usage{
bench_mark(data, row_num = 4000, resamplings = rsmps("bootstrap", repeats = 5))
}
\arguments{
\item{data}{A tibble.All are numeric except the first column is a factor.}

\item{row_num}{Number of samples used.}

\item{resamplings}{R6/Resampling.Resampling method.}
}
\description{
Compare classification models with small samples.
}
\examples{
data(ProcData)
}
