% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{col_split}
\alias{col_split}
\title{Cut the string of ID column character by character and divide it into multiple columns.}
\usage{
col_split(data, cores = NULL, pad_len = 10)
}
\arguments{
\item{data}{Dataframe(tibble) to be split.}

\item{cores}{Int.The num of cores to allocate for computing.}

\item{pad_len}{The length of longest id, i.e. the maxlength.}
}
\value{
A tibble with pad_len+1 column.
}
\description{
Cut the string of ID column character by character and divide it into multiple columns.
}
\examples{
data(ID_sample)
col_split(ID_sample,cores = 1,pad_len = 10)
}
