% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{data_clean}
\alias{data_clean}
\title{Reshape data and delete meaningless rows.}
\usage{
data_clean(data, cols = everything(), placeholder = c("-"))
}
\arguments{
\item{data}{A dataframe or tibble or data.table or matrix. Names of the column will be regard as the class of ID included in column.}

\item{cols}{Character vectors. Columns of \code{data} that contain the IDs}

\item{placeholder}{Character vectors. IDs included in \code{placeholder} will be omitted.}
}
\value{
A tibble with two columns("ID" and "class")
}
\description{
Reshape data and delete meaningless rows.
}
\examples{
data <- tibble::tibble(
"class1" = c("A","B","C","D"),
'class2' = c("E","F","G","H"),
'class3' = c("L","M","-","O")
)
data_clean(data)
}
