% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.R
\docType{data}
\name{Example}
\alias{Example}
\title{ID example dataset.}
\format{
A tibble with 5000 rows and 2 variables.
\describe{
\item{ID}{A identifier character.}
\item{class}{The database the ID belongs to.}
}
}
\usage{
Example
}
\description{
ID example dataset.
}
\keyword{datasets}
