% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GETID.R
\name{GETID}
\alias{GETID}
\title{Get ID data from \code{Biomart} database use \code{attributes}.}
\usage{
GETID(attributes, Ensembl = NULL)
}
\arguments{
\item{attributes}{A dataframe.The information we want to retrieve.Use \code{GET_ID_attributes} to hava try.}

\item{Ensembl}{\code{Mart} object from \code{biomaRt}.}
}
\value{
A \code{tibble} dataframe.
}
\description{
Get ID data from \code{Biomart} database use \code{attributes}.
}
\examples{
data(attributes)
GETID(attributes[1:3, ])
}
