% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelExplain.R
\name{htmp_confu}
\alias{htmp_confu}
\title{Plot heatmap for result confusion matrix.}
\usage{
htmp_confu(table, name, filepath = NULL)
}
\arguments{
\item{table}{A table.}

\item{name}{Model names.}

\item{filepath}{File path the plot to save.Default NULL.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot heatmap for result confusion matrix.
}
