% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID_sample.R
\docType{data}
\name{ID_sample}
\alias{ID_sample}
\title{ID dataset for testing.}
\format{
A tibble with 5000 rows and 2 variables.
\describe{
\item{ID}{A identifier character.}
\item{class}{The database the ID belongs to.}
}
}
\usage{
ID_sample
}
\description{
ID dataset for testing.
}
\keyword{datasets}
