% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{mi}
\alias{mi}
\title{A wrapper function that executes MantaID workflow.}
\usage{
mi(
  mi_data_allID,
  cores = NULL,
  levels = c("*", 0:9, letters, LETTERS, "_", ".", "-", " ", "/", "\\\\", ":"),
  ratio = 0.3,
  para_blc = FALSE,
  model_path = NULL,
  batch_size = 128,
  epochs = 64,
  validation_split = 0.3,
  graph_path = NULL
)
}
\arguments{
\item{mi_data_allID}{IDs raw data.}

\item{cores}{The number of cores used when balancing data.}

\item{levels}{The vector that includes all the single character occurred in IDs.}

\item{ratio}{The ratio of test set.}

\item{para_blc}{A logical value whether using parallel computing when balancing data.}

\item{model_path}{The path to save models.}

\item{batch_size}{The batch size of deep learning model fitting.}

\item{epochs}{The epochs of deep learning model fitting.}

\item{validation_split}{The validation ratio of deep learning model fitting.}

\item{graph_path}{The path to save graphs.}
}
\value{
The list of models and graphs.
}
\description{
A wrapper function that executes MantaID workflow.
}
