% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_balance_data.R
\name{mi_balance_data}
\alias{mi_balance_data}
\title{Data balance.
Most classes adopt random undersampling, while a few classes adopt smote method to oversample to obtain relatively balanced data;}
\usage{
mi_balance_data(data, ratio = 0.3, parallel = FALSE)
}
\arguments{
\item{data}{A data frame. Except class column, all are numeric types.}

\item{ratio}{Numeric between 0 and 1. The percent of test set split from data.}

\item{parallel}{Logical.}
}
\value{
A list contain train set and test set.
}
\description{
Data balance.
Most classes adopt random undersampling, while a few classes adopt smote method to oversample to obtain relatively balanced data;
}
\examples{
library(dplyr)
data = rename(iris,class =Species)
mi_balance_data(data)
}
