% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_data_procID.R
\docType{data}
\name{mi_data_procID}
\alias{mi_data_procID}
\title{Processed ID data.}
\format{
A tibble dataframe with 5000 rows and 21 variables.
\describe{
\item{pos1 to pos20}{Splited ID.}
\item{class}{The databases that ID belongs to.}
}
}
\usage{
mi_data_procID
}
\description{
Processed ID data.
}
\keyword{datasets}
