% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_get_importance.R
\name{mi_get_importance}
\alias{mi_get_importance}
\title{Plot the bar plot for feature importance.}
\usage{
mi_get_importance(data)
}
\arguments{
\item{data}{A table.}
}
\value{
A bar plot.
}
\description{
Plot the bar plot for feature importance.
}
