% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_get_padlen.R
\name{mi_get_padlen}
\alias{mi_get_padlen}
\title{Get max length of ID data.}
\usage{
mi_get_padlen(data)
}
\arguments{
\item{data}{A dataframe.}
}
\value{
A int.
}
\description{
Get max length of ID data.
}
\examples{
data(mi_data_rawID)
mi_get_padlen(mi_data_rawID)
}
