% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_plot_heatmap.R
\name{mi_plot_heatmap}
\alias{mi_plot_heatmap}
\title{Plot heatmap for result confusion matrix.}
\usage{
mi_plot_heatmap(table, name = NULL, filepath = NULL)
}
\arguments{
\item{table}{A table.}

\item{name}{Model names.}

\item{filepath}{File path the plot to save.Default NULL.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot heatmap for result confusion matrix.
}
