% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_predict_new.R
\name{mi_predict_new}
\alias{mi_predict_new}
\title{Predict new data with a trained learner.}
\usage{
mi_predict_new(data, result, ifnet = F)
}
\arguments{
\item{data}{A dataframe.}

\item{result}{The result object from a previous training.}

\item{ifnet}{A boolean indicating if a neural network is used for prediction.}
}
\value{
A data frame that contains features and 'predict' class.
}
\description{
Predict new data with a trained learner.
}
