% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_run_bmr.R
\name{mi_run_bmr}
\alias{mi_run_bmr}
\title{Compare classification models with small samples.}
\usage{
mi_run_bmr(data, row_num = 1000, resamplings = rsmps("cv", folds = 10))
}
\arguments{
\item{data}{A tibble.All are numeric except the first column is a factor.}

\item{row_num}{Number of samples used.}

\item{resamplings}{R6/Resampling.Resampling method.}
}
\value{
A list of R6 class of benchmark results and scores of test set.
examples
data(mi_data_procID)
mi_run_bmr(mi_data_procID)
}
\description{
Compare classification models with small samples.
}
