% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_split_col.R
\name{mi_split_col}
\alias{mi_split_col}
\title{Cut the string of ID column character by character and divide it into multiple columns.}
\usage{
mi_split_col(data, cores = NULL, pad_len = 10)
}
\arguments{
\item{data}{Dataframe(tibble) to be split.}

\item{cores}{Int.The num of cores to allocate for computing.}

\item{pad_len}{The length of longest id, i.e. the maxlength.}
}
\value{
A tibble with pad_len+1 column.
}
\description{
Cut the string of ID column character by character and divide it into multiple columns.
}
