% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_split_str.R
\name{mi_split_str}
\alias{mi_split_str}
\title{Split the string into individual characters and complete the character vector to the maximum length.}
\usage{
mi_split_str(str, pad_len)
}
\arguments{
\item{str}{The string to be splited.}

\item{pad_len}{The length of longest id, i.e. the maxlength.}
}
\value{
Splited character vector.
}
\description{
Split the string into individual characters and complete the character vector to the maximum length.
}
\examples{
string_test = "Good Job"
length = 15
mi_split_str(string_test,length)
}
