% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_train_rg.R
\name{mi_train_rg}
\alias{mi_train_rg}
\title{Random Forest Model Training.}
\usage{
mi_train_rg(train, test, measure = msr("classif.acc"), instance = NULL)
}
\arguments{
\item{train}{A dataframe.}

\item{test}{A dataframe.}

\item{measure}{Model evaluation method.}

\item{instance}{A tuner.}
}
\value{
A list of learner for predict and predict result of test set.
}
\description{
Random Forest Model Training.
}
