% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_train_rp.R
\name{mi_train_rp}
\alias{mi_train_rp}
\title{Classification tree model training.}
\usage{
mi_train_rp(train, test, measure = msr("classif.acc"), instance = NULL)
}
\arguments{
\item{train}{A dataframe.}

\item{test}{A dataframe.}

\item{measure}{Model evaluation method.Use \code{mlr_measures} and \code{msr()} to view and choose metrics.}

\item{instance}{A tuner.}
}
\value{
A list of learner for predict and predict result of test set.
}
\description{
Classification tree model training.
}
