% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_unify_mod.R
\name{mi_unify_mod}
\alias{mi_unify_mod}
\title{Predict with four models and unify results by the sub-model's specificity score to the four possible classes.}
\usage{
mi_unify_mod(
  data,
  col_id,
  result_rg,
  result_rp,
  result_xgb,
  result_BP,
  c_value = 0.75,
  pad_len = 30
)
}
\arguments{
\item{data}{A dataframe contains the ID column.}

\item{col_id}{The name of ID column.}

\item{result_rg}{The result from the Random Forest model.}

\item{result_rp}{The result from the Decision Tree model.}

\item{result_xgb}{The result from the XGBoost model.}

\item{result_BP}{The result from the Backpropagation Neural Network model.}

\item{c_value}{A numeric value used in the final prediction calculation.}

\item{pad_len}{The length to pad the ID characters to.}
}
\value{
A dataframe.
}
\description{
Predict with four models and unify results by the sub-model's specificity score to the four possible classes.
}
