% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelTrain.R
\name{model_rg}
\alias{model_rg}
\title{Random Forest Model Training.}
\usage{
model_rg(train, test, measure = msr("classif.acc"))
}
\arguments{
\item{train}{A dataframe.}

\item{test}{A dataframe.}

\item{measure}{Model evaluation method.}
}
\description{
Random Forest Model Training.
}
