% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelTrain.R
\name{model_rp}
\alias{model_rp}
\title{Classification tree model training.}
\usage{
model_rp(train, test, measure = msr("classif.acc"))
}
\arguments{
\item{train}{A dataframe.}

\item{test}{A dataframe.}

\item{measure}{Model evaluation method.Use \code{mlr_measures} and \code{msr()} to view and choose metrics.}
}
\description{
Classification tree model training.
}
