% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelTrain.R
\name{model_xgboost}
\alias{model_xgboost}
\title{Xgboost model training}
\usage{
model_xgboost(train, test, measure = msr("classif.acc"))
}
\arguments{
\item{train}{A dataframe.}

\item{test}{A dataframe.}

\item{measure}{Model evaluation method.}
}
\description{
Xgboost model training
}
