% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelExplain.R
\name{mstk_prd}
\alias{mstk_prd}
\title{Observe the distribution of the false response of test set.}
\usage{
mstk_prd(predict)
}
\arguments{
\item{predict}{A R6 class \code{PredictionClassif}.}
}
\value{
A tibble data frame that records the number of wrong predictions for each category ID;
}
\description{
Observe the distribution of the false response of test set.
}
