% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_data_attributes.R
\docType{data}
\name{mt_data_attributes}
\alias{mt_data_attributes}
\title{ID-related datasets in biomart.}
\format{
A dataframe with 65 variables and 3 variables.
\describe{
\item{name}{The name of dataset.}
\item{description}{Description of dataset.}
\item{page}{collection of attributes.}
}
}
\usage{
mt_data_attributes
}
\description{
ID-related datasets in biomart.
}
\keyword{datasets}
