% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_data_rawID.R
\docType{data}
\name{mt_data_rawID}
\alias{mt_data_rawID}
\title{ID dataset for testing.}
\format{
A tibble with 5000 rows and 2 variables.
\describe{
\item{ID}{A identifier character.}
\item{class}{The database the ID belongs to.}
}
}
\usage{
mt_data_rawID
}
\description{
ID dataset for testing.
}
\keyword{datasets}
