% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_get_confusion.R
\name{mt_get_confusion}
\alias{mt_get_confusion}
\title{Compute the confusion matrix for the predict result.}
\usage{
mt_get_confusion(result_list, ifnet = F)
}
\arguments{
\item{result_list}{A list return from model training functions.}

\item{ifnet}{Logical.Whether the data is obtained by a deep learning model.}
}
\value{
A \code{confusionMatrix} object.
}
\description{
Compute the confusion matrix for the predict result.
}
