% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_get_ID.R
\name{mt_get_ID}
\alias{mt_get_ID}
\title{Get ID data from \code{Biomart} database use \code{attributes}.}
\usage{
mt_get_ID(
  attributes,
  biomart = "genes",
  dataset = "hsapiens_gene_ensembl",
  mirror = "asia"
)
}
\arguments{
\item{attributes}{A dataframe.The information we want to retrieve.Use \code{mt_get_ID_attr} to hava try.}

\item{biomart}{BioMart database name you want to connect to.Use \code{biomaRt::listEnsembl} to retrieve the possible database names.}

\item{dataset}{Datasets of the selected BioMart database.}

\item{mirror}{Specify an Ensembl mirror to connect to.}
}
\value{
A \code{tibble} dataframe.
}
\description{
Get ID data from \code{Biomart} database use \code{attributes}.
}
\examples{
data(mt_data_attributes)
mt_get_ID(mt_data_attributes[1:3, ])
}
