% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_get_ID_attr.R
\name{mt_get_ID_attr}
\alias{mt_get_ID_attr}
\title{Get ID attributes from \code{Biomart} database.}
\usage{
mt_get_ID_attr(
  biomart = "genes",
  dataset = "hsapiens_gene_ensembl",
  mirror = "asia"
)
}
\arguments{
\item{biomart}{BioMart database name you want to connect to.Use \code{biomaRt::listEnsembl} to retrieve the possible database names.}

\item{dataset}{Datasets of the selected BioMart database.}

\item{mirror}{Specify an Ensembl mirror to connect to.}
}
\value{
A dataframe.
}
\description{
Get ID attributes from \code{Biomart} database.
}
