% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_get_padlen.R
\name{mt_get_padlen}
\alias{mt_get_padlen}
\title{Get max length of ID data.}
\usage{
mt_get_padlen(data)
}
\arguments{
\item{data}{A dataframe.}
}
\value{
A int.
}
\description{
Get max length of ID data.
}
