% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_predict_new.R
\name{mt_predict_new}
\alias{mt_predict_new}
\title{Predict new data with trained learner.}
\usage{
mt_predict_new(data, learner)
}
\arguments{
\item{data}{A dataframe.}

\item{learner}{A R6 class object.}
}
\description{
Predict new data with trained learner.
}
