% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_to_numer.R
\name{mt_to_numer}
\alias{mt_to_numer}
\title{Convert data to numeric, and for ID column convert with fixed levels.}
\usage{
mt_to_numer(
  data,
  levels = c("*", 0:9, letters, LETTERS, "_", ".", "-", " ", "/", "\\\\", ":")
)
}
\arguments{
\item{data}{A tibble with n position column(pos1,pos2,...) and class column.}

\item{levels}{Characters accommodated in IDs.}
}
\description{
Convert data to numeric, and for ID column convert with fixed levels.
}
\examples{
data(mt_data_rawID)
str(mt_split_col(mt_data_rawID,cores = 1,pad_len = 10))
str(mt_to_numer(mt_split_col(mt_data_rawID,cores = 1,pad_len = 10)))
}
