% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_train_xgb.R
\name{mt_train_xgb}
\alias{mt_train_xgb}
\title{Xgboost model training}
\usage{
mt_train_xgb(train, test, measure = msr("classif.acc"))
}
\arguments{
\item{train}{A dataframe.}

\item{test}{A dataframe.}

\item{measure}{Model evaluation method.}
}
\description{
Xgboost model training
}
