% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_tune_rg.R
\name{mt_tune_rg}
\alias{mt_tune_rg}
\title{Tune Random Forest model by hyperband.}
\usage{
mt_tune_rg(
  data,
  resampling = rsmp("cv", folds = 5),
  measure = msr("classif.acc"),
  eta = 3
)
}
\arguments{
\item{data}{A tibble.All are numeric except the first column is a factor.}

\item{resampling}{R6/Resampling.}

\item{measure}{Model evaluation method.Use \code{mlr_measures} and \code{msr()} to view and choose metrics.}

\item{eta}{The percent parameter configurations discarded.}
}
\value{
A list of tuning instance and stage plot.
}
\description{
Tune Random Forest model by hyperband.
}
