% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelExplain.R
\name{prd_new}
\alias{prd_new}
\title{Predict new data with trained learner.}
\usage{
prd_new(data, learner)
}
\arguments{
\item{data}{A dataframe.}

\item{learner}{A R6 class object.}
}
\description{
Predict new data with trained learner.
}
