% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcData.R
\docType{data}
\name{ProcData}
\alias{ProcData}
\title{Processed ID data.}
\format{
A tibble dataframe with 5000 rows and 21 variables.
\describe{
\item{pos1 to pos20}{Splited ID.}
\item{class}{The databases that ID belongs to.}
}
}
\usage{
ProcData
}
\description{
Processed ID data.
}
\keyword{datasets}
