% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{string_split}
\alias{string_split}
\title{Split the string into individual characters and complete the character vector to the maximum length.}
\usage{
string_split(str, pad_len)
}
\arguments{
\item{str}{The string to be splited.}

\item{pad_len}{The length of longest id, i.e. the maxlength.}
}
\value{
Splited character vector.
}
\description{
Split the string into individual characters and complete the character vector to the maximum length.
}
\examples{
string_test = "Good Job"
length = 15
string_split(string_test,length)
}
