% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{to_factor}
\alias{to_factor}
\title{Convert data to factor type, and for ID column convert with fixed levels.}
\usage{
to_factor(
  data,
  levels = c("*", 0:9, letters, LETTERS, "_", ".", "-", " ", "/", "\\\\", ":")
)
}
\arguments{
\item{data}{A tibble with n position column(pos1,pos2,...) and class column.}

\item{levels}{Characters accommodated in IDs.
data(ID_sample)
str(col_split(ID_sample,cores = 1,pad_len = 10))
str(to_factor(col_split(ID_sample,cores = 1,pad_len = 10)))}
}
\description{
Convert data to factor type, and for ID column convert with fixed levels.
}
