% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_plot_cor.R
\name{mi_plot_cor}
\alias{mi_plot_cor}
\title{Plot correlation heatmap.}
\usage{
mi_plot_cor(data, cls = "class")
}
\arguments{
\item{data}{Data frame that including IDs' position features.}

\item{cls}{The name of the class column.}
}
\value{
A heatmap.
}
\description{
Plot correlation heatmap.
}
\examples{
data(mi_data_procID)
data_num <- mi_to_numer(mi_data_procID)
mi_plot_cor(data_num)
}
