% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_to_numer.R
\name{mi_to_numer}
\alias{mi_to_numer}
\title{Convert data to numeric, and for ID column convert with fixed levels.}
\usage{
mi_to_numer(
  data,
  levels = c("*", 0:9, letters, LETTERS, "_", ".", "-", " ", "/", "\\\\", ":")
)
}
\arguments{
\item{data}{A tibble with n position column(pos1,pos2,...) and class column.}

\item{levels}{Characters accommodated in IDs.}
}
\description{
Convert data to numeric, and for ID column convert with fixed levels.
}
\examples{
data(mi_data_rawID)
str(mi_split_col(mi_data_rawID,cores = 1,pad_len = 10))
str(mi_to_numer(mi_split_col(mi_data_rawID,cores = 1,pad_len = 10)))
}
