% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_train_rg.R
\name{mi_train_rg}
\alias{mi_train_rg}
\title{Random Forest Model Training.}
\usage{
mi_train_rg(train, test, measure = msr("classif.acc"))
}
\arguments{
\item{train}{A dataframe.}

\item{test}{A dataframe.}

\item{measure}{Model evaluation method.}
}
\description{
Random Forest Model Training.
}
