% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_unify_mod.R
\name{mi_unify_mod}
\alias{mi_unify_mod}
\title{Predict with four models and unify results by the sub-model's specificity score to the four possible classes.}
\usage{
mi_unify_mod(
  data,
  col_id,
  result_rg,
  result_rp,
  result_xgb,
  result_BP,
  c_value = 0.75,
  pad_len = 30
)
}
\arguments{
\item{data}{A dataframe contains the ID column.}

\item{col_id}{The name of ID column.}
}
\value{
A dataframe.
}
\description{
Predict with four models and unify results by the sub-model's specificity score to the four possible classes.
}
